refactor csp build
